ads-integration-service/requirements.txt
mysql~=0.0.3
mysql-connector-python
requests~=2.32.3
google~=3.0.0
pika~=1.3.1
python-dotenv~=1.0.0
facebook_business
google-ads
utils

apsscheduler